Reorganise packages and finish splitting Graph and CQL classes
