+++ src/main/scala/com/datastax/gatling/plugin/request/CqlRequestActionBuilder.scala
@@ -7,17 +7,18 @@
 package com.datastax.gatling.plugin.request
 
 import com.datastax.gatling.plugin.DseProtocol
+import com.datastax.gatling.plugin.model.DseCqlAttributes
 import io.gatling.core.action.Action
 import io.gatling.core.action.builder.ActionBuilder
 import io.gatling.core.structure.ScenarioContext
 import io.gatling.core.util.NameGen
 
-class DseCqlRequestActionBuilder(dseAttributes: DseCqlAttributes) extends ActionBuilder with
+class CqlRequestActionBuilder(dseAttributes: DseCqlAttributes) extends ActionBuilder with
   NameGen {
 
   def build(ctx: ScenarioContext, next: Action): Action = {
     val dseComponents = ctx.protocolComponentsRegistry.components(DseProtocol.DseProtocolKey)
-    new DseCqlRequestAction(
+    new CqlRequestAction(
       dseAttributes.tag,
       next,
       ctx.system,
@@ -29,20 +30,4 @@ class DseCqlRequestActionBuilder(dseAttributes: DseCqlAttributes) extends Action
       dseComponents.gatlingTimingSource)
   }
 }
-class DseGraphRequestActionBuilder(dseAttributes: DseGraphAttributes) extends ActionBuilder with
-  NameGen {
 
-  def build(ctx: ScenarioContext, next: Action): Action = {
-    val dseComponents = ctx.protocolComponentsRegistry.components(DseProtocol.DseProtocolKey)
-    new DseGraphRequestAction(
-      dseAttributes.tag,
-      next,
-      ctx.system,
-      ctx.coreComponents.statsEngine,
-      dseComponents.dseProtocol,
-      dseAttributes,
-      dseComponents.metricsLogger,
-      dseComponents.dseRequestsRouter,
-      dseComponents.gatlingTimingSource)
-  }
-}
