+++ src/main/scala/com/datastax/gatling/plugin/request/DseRequestActor.scala
@@ -17,8 +17,8 @@ import io.gatling.core.session.Session
 import scala.concurrent.ExecutionException
 import scala.util.{Failure, Success, Try}
 
-case class SendCqlQuery(dseRequestAction: DseCqlRequestAction, session: Session)
-case class SendGraphQuery(dseRequestAction: DseGraphRequestAction, session: Session)
+case class SendCqlQuery(dseRequestAction: CqlRequestAction, session: Session)
+case class SendGraphQuery(dseRequestAction: GraphRequestAction, session: Session)
 
 case class RecordResult[T](t: Try[T], callback: FutureCallback[T])
